uniformize semester types
